+++ repository/get_session_request.go
@@ -1,0 +1,11 @@
+package repository
+
+import "github.com/google/uuid"
+
+type GetSessionByIDRequest struct {
+	ID uuid.UUID
+}
+
+type GetSessionByUserIDRequest struct {
+	UserID int64
+}
